requirements.txt
aiogram==3.18.0
redis==5.2.1
python-dotenv==1.0.1
aiohttp==3.11.13